+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * A new heading can now be set with teleport commands.
 * Added zone objects with the `ZoneObject` class.
 * Replace a bunch of `CustomMessage` properties with `WorldCommand` ones.
+* Added a command to be run when walking in a box.
 
 ## [0.13.0]